cloud/webPage/requirements.txt
functions-framework==3.*
google-cloud-firestore
Flask
matplotlib